fix: [BE] 코드 리뷰 반영 및 결함 수정 (#56)

+++ storage/db-core/src/main/java/com/coffiness/calfit/storage/db/core/recruitment/RecruitmentInterviewerRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RecruitmentInterviewerRepository
-    extends JpaRepository<RecruitmentInterviewerEntity, Integer> {
+    extends JpaRepository<RecruitmentInterviewerEntity, Long> {
   List<RecruitmentInterviewerEntity> findByRecruitmentIdIn(List<Long> recruitmentIds);
 
   List<RecruitmentInterviewerEntity> findByRecruitmentId(Long recruitmentId);
